requirements.txt
Django==1.8.5
pyserial==2.7
PyYAML==3.11
RPi.GPIO==0.5.11
tornado==4.2.1
wheel==0.24.0
